+++ src/screens/app/CreateMemeScreen.tsx
@@ -171,10 +171,11 @@ const CreateMemeScreen: CreateMemeScreen = ({ navigation }) => {
 					style={styles.input}
 					value={title}
 					onChangeText={setTitle}
+					autoCapitalize="none"
 				/>
 
 				<Text style={styles.label}>Category</Text>
-				<DropDownPicker
+        <DropDownPicker
 					open={open}
 					value={selectedCategory}
 					items={items}
@@ -214,7 +215,7 @@ const styles = StyleSheet.create({
 	container: {
 		flex: 1,
 		padding: 16,
-		backgroundColor: 'rgba(0,0,0,0.9)',
+		backgroundColor: 'rgba(0,0,0,0.8)',
 	},
 	formContainer: {
 		padding: 16,